Code supposément compatible avec l'interface graphique
